splitt MicroCards fra MicroCard
